scraper/requirements.txt
playwright==1.52.0
beautifulsoup4==4.13.4
sqlalchemy==2.0.41
psycopg2-binary==2.9.10
python-dotenv==1.1.0
apscheduler==3.11.0
